Separate abstract and sync unit tests
